requirements.txt
fastapi
uvicorn
pydantic
pywebview
